+++ src/utils/dnd.ts
@@ -65,44 +65,6 @@ function findFirstDroppable(collisions: Collision[] | null) {
   return collisions?.find(({ data }) => !!data?.droppableContainer);
 }
 
-function isTransitionAllowed<T>(
-  stateWorkflow: WorkflowTransitionsResponse | undefined,
-  { initial, containerId }: DragUpdateEventArgs<T>
-) {
-  return process.env.NEXT_PUBLIC_FEATURE_PROJECT_USERS_WORKFLOW === "true"
-    ? !!(
-        stateWorkflow &&
-        initial?.containerId &&
-        containerId &&
-        (initial.containerId === containerId ||
-          stateWorkflow[initial.containerId].includes(containerId))
-      )
-    : true;
-}
-
-function getItemsByTransitions<
-  T extends WithModelState<{ id: UniqueIdentifier }>,
->(
-  stateWorkflow: WorkflowTransitionsResponse | undefined,
-  data: T[] | undefined
-): DndItems<T> | null {
-  if (stateWorkflow && data?.length) {
-    const items: DndItems<T> = {};
-
-    Object.keys(stateWorkflow).forEach((key: string) => {
-      items[key] = [];
-    });
-
-    data.forEach(item => {
-      items[item.model_state.state.slug].push(item);
-    });
-
-    return items;
-  }
-
-  return null;
-}
-
 export {
   findItemInContainer,
   findContainer,
@@ -111,6 +73,4 @@ export {
   findFirstDroppable,
   pruneItem,
   findDroppables,
-  isTransitionAllowed,
-  getItemsByTransitions,
 };
